backend/foodgram_backend/requirements.txt
asgiref==3.7.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2.19
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.2.0
drf-extra-fields==3.5.0
filetype==1.2.0
flake8==6.0.0
fonttools==4.40.0
html5lib==1.1
idna==3.4
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.5.0
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.6.0
pyflakes==3.0.1
Pygments==2.15.1
pyphen==0.14.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
weasyprint==59.0
webencodings==0.5.1
zopfli==0.2.2